Limit activator to sending activation and db request
